Replace dummy data with ES finder for risk score resolver

See merge request https://gitlab.com/gitlab-org/gitlab/-/merge_requests/211126

Merged-by: Subashis  Chakraborty <schakraborty@gitlab.com>
Approved-by: Charlie Kroon <ckroon@gitlab.com>
Approved-by: Gregory Havenga <11164960-ghavenga@users.noreply.gitlab.com>

+++ ee/app/graphql/types/security/risk_score_type.rb
@@ -6,7 +6,7 @@ module Types
       graphql_name 'RiskScore'
       description 'Total risk score information'
 
-      field :score, GraphQL::Types::Int,
+      field :score, GraphQL::Types::Float,
         null: false,
         experiment: { milestone: '18.4' },
         description: 'Overall risk score.'
